+++ docs/source/release_notes/release-notes.rst
@@ -59,6 +59,11 @@ develop
          - Changes the default scrubber behavior to aggressive scrub (synchronous with scrub request).
            (`Pull Request <https://github.com/palantir/atlasdb/pull/3009>`__)
 
+    *    - |improved|
+         - Remove a round trip from read only transactions
+           not involving thoroughly swept tables.
+           (`Pull Request <https://github.com/palantir/atlasdb/pull/3020>`__)
+
 =======
 v0.77.0
 =======
@@ -93,7 +98,6 @@ v0.77.0
     *    - |fixed|
          - Fix ``SnapshotTransaction#getRows`` to apply ``ColumnSelection`` when there are local writes.
            (`Pull Request <https://github.com/palantir/atlasdb/pull/3008>`__)
-
 
 =======
 v0.76.0